Added Catnip Powerup!
Cleaned up a bit of code and added the Awesome component, which so far is only used on catnip, which when hit causes a temporary speed boost. Also implemented a max energy level of 1000 and made food reduce energy in order to prevent infinite games. Energy is now meant to represent craziness, food reduces crazyness, toys and bugs increase crazyness.