Added Flyer Component and Slowed game speed.
The Flyer component allows for a knockable to follow an oval movement path. Created a Fly enemy which utilizes the Flyer component and flys around the stage. Slowed the game speed to increase user control and ease debugging.

Fixed a few issues with entity animations and other small tweaks I'm forgetting.